Added separate trail shader
Let me know if this doesnt work for anyone! Essentially replaced lab7
shader with tav and trail shader.  Trail is an exact copy of lab7 and
tav is a copy minus the fog stuff.  Basically just made copies of all
the handles.... One of these days i'll clean it up.